backend/requirements.txt
fastapi==0.115.12
uvicorn[standard]==0.34.0
python-dotenv==1.0.1
httpx==0.27.0
google-generativeai
radon==6.0.1
bandit==1.7.9
pylint==3.2.2
python-multipart==0.0.9
pypdf==5.4.0
